python/twitter_bot/requirements.txt
gql
Pillow
requests_toolbelt
tweepy
Web3